pyproject.toml
[project]
name = "ai-mangaka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.7.0",
    "matplotlib>=3.9.0",
    "pillow>=11.0.0",
    "requests>=2.32.0",
]
